% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_url}
\alias{assert_url}
\title{Check if an argument is a URL}
\usage{
assert_url(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
If the check is successful, x will be returned invisibly.
}
\description{
Check if an argument is a URL
}
\examples{
url <- "https://github.com/KoderKow/pathr"

assert_url(url)

# The following throws an error
url <- "ww.exampl.co/this/will/fail"

try(assert_url(url))
}
