% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_functions.R
\name{canvas}
\alias{canvas}
\title{Canvas generator}
\usage{
canvas(x_max, y_max, col_width = 1, border_pad = 1)
}
\arguments{
\item{x_max}{Maximum canvas width of maze for \eqn{x}-axis.}

\item{y_max}{Maximum canvas width of maze for \eqn{y}-axis.}

\item{col_width}{Size of columns for canvas. Rounds to nearest whole number.}

\item{border_pad}{Uniform margin width for canvas.}
}
\description{
Generates the overall canvas with respective dimensions for
maze generation.
}
