% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_functions.R
\name{cell}
\alias{cell}
\title{Cell constructor}
\usage{
cell(x, y, cell_width, walls = c(TRUE, TRUE, TRUE, TRUE))
}
\arguments{
\item{x}{Cell position on coordinate \eqn{x}.}

\item{y}{Cell position on coordinate \eqn{y}.}

\item{cell_width}{Overall width of cell to be drawn.}

\item{walls}{Boolean vector to indicate whether wall of cell should be
drawn.}
}
\description{
Constructs a 4-sided polygon at a given \eqn{x} and \eqn{y}
coordinates.
}
\details{
\code{walls} will need to be a boolean vector. To import this
parameter, the order is \code{bottom, left, top, and right}. If element
is set to \code{FALSE}, no wall will be drawn.
}
