% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copy.R
\name{copy}
\alias{copy}
\title{Pipeable Clipboard Copying}
\usage{
copy(x, ...)
}
\arguments{
\item{x}{An object to be written to the system clipboard.}

\item{\ldots}{Other arguments passed to \code{\link[clipr]{write_clip}}.}
}
\value{
Returns \code{x} but attempts to copy content to the clipboard.
}
\description{
This convenience function wraps \code{\link[clipr]{write_clip}} to copy the
previous output in the chain to the clipboard.
}
\examples{
\dontrun{
library(magrittr)
files \%>\%
    parse_path() \%>\%
    swap_regex("^(P|R)", "oops") \%>\%
    copy()
}
}

