% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_database.R
\name{create_database}
\alias{create_database}
\title{Create a learning database}
\usage{
create_database()
}
\value{
List of tibbles.
}
\description{
Function creating an empty learning database.
}
\author{
Nicolas Mangin
}
