% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_path.R
\name{create_path}
\alias{create_path}
\title{Create a learning path}
\usage{
create_path(activities, connections)
}
\arguments{
\item{activities}{Tibble. List of pre-filtered activities with their attributes to be organized in a learning path.}

\item{connections}{Tibble. List of edges between learning outcomes.}
}
\value{
Tibble. A list of link between activities (origin, destination, and condition)
}
\description{
Function creating a learning path from selected activities and connections between learning outcomes.
}
\author{
Nicolas Mangin
}
