% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_path_ui.R
\name{design_path_ui}
\alias{design_path_ui}
\title{Design learning paths}
\usage{
design_path_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified document in the original documents folder.
}
\description{
Module facilitating the design and visualization or a complete learning experience.
}
\author{
Nicolas Mangin
}
