% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_path.R
\name{expand_path}
\alias{expand_path}
\title{Expand Tilde Path}
\usage{
expand_path(x)
}
\arguments{
\item{x}{A tilde prefixed path.}
}
\value{
Returns a vector of expanded paths.
}
\description{
A wrapper for \code{\link[base]{path.expand}} that expands a tilde prefixed
path, replacing the leading tilde by the user's home directory (if defined on
that platform).
}
\examples{
library(magrittr)
file_path("~/mydir", paste0("file", 1:2, ".pdf")) \%>\%
    expand_path()
}
\seealso{
\code{\link[base]{file.path}}
}

