% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path.R
\name{file_path}
\alias{file_path}
\title{Construct Path to File}
\usage{
file_path(...)
}
\arguments{
\item{\ldots}{Atomic vectors of elements to construct a path from.}
}
\value{
Returns a vector of paths.
}
\description{
A wrapper for \code{\link[base]{file.path}} that normalizes the path separator
between elements.
}
\examples{
file_path("mydir", paste0("file", 1:2, ".pdf"))
}
\seealso{
\code{\link[base]{file.path}}
}
