% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathr-package.R
\docType{data}
\name{files}
\alias{files}
\title{A Vector of File Paths}
\format{A character vector with 203 elements}
\usage{
data(files)
}
\description{
A dataset containing a character vector of file paths.
}
\keyword{datasets}

