% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.paths.R
\name{find.paths}
\alias{find.paths}
\title{Path detection}
\usage{
find.paths(pts, distal.pts, proximal.z, min.range, max.range)
}
\arguments{
\item{pts}{and object created using the function read.cloud().}

\item{distal.pts}{an object created using the function get.distal.pts().}

\item{proximal.z}{value, paths' end point height (typically, 0.0 or breast height).}

\item{min.range}{value, minimal search radius.}

\item{max.range}{value, maximal search radius.}
}
\value{
a list including the created paths.
}
\description{
Detect paths from distal points to a proximal point. Plagiotropic path: includes at least one segment <=0.0 degree angle; orthotropic path: all segment angles >0.0 degree; 0 degree indicates a perfectly horizontal path segment.
}
\examples{
# read point cloud
pts <- read.cloud(file = "C:/.../point_cloud_file_name.txt", sep = " ", plot = TRUE) 
# get distal points
distal.pts <- get.distal.pts(pts, min.z = 1.3, plot = TRUE)
# get paths
paths <- find.paths(pts, distal.pts, proximal.z = 0.0, min.range  = .02, max.range  = .08)

}
