% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/front.R
\name{front}
\alias{back}
\alias{front}
\title{Extract the First/Last N Elements of a Path}
\usage{
front(x, n = -1, collapse = TRUE, ...)

back(x, n = 1, collapse = TRUE, ...)
}
\arguments{
\item{x}{A \code{parse_path} object.}

\item{n}{A single integer. If positive, size for the resulting object: number
of elements for a vector. If negative, all but the \code{n} last/first number
of elements of \code{x}.}

\item{collapse}{logical.  If \code{TRUE} the path elements will be collapsed
with the \code{"/"} separator.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector (either \code{atomic} or \code{\link[base]{list}} of
\code{character} path elements) of sub-paths.
}
\description{
\code{front} - Extract the first n elements of a path.  Similar to
\code{\link[package]{head}} (uses \code{\link[package]{head}} under the hood).
Compare the defualt settings to \code{\link[base]{basename}} and
\code{\link[base]{dirname}}.

\code{back} - Extract the first n elements of a path.  Similar to
\code{\link[package]{tail}} (uses \code{\link[package]{tail}} under the hood).
}
\examples{
front(list(1:8), 3)
front(list(1:8), -3)

back(list(1:8), 3)
back(list(1:8), -3)

library(magrittr)
files \%>\%
    parse_path() \%>\%
    front()

files \%>\%
    parse_path() \%>\%
    front(3)

files \%>\%
    parse_path() \%>\%
    back()

files \%>\%
    parse_path() \%>\%
    back(3)

## Compare to basename/dirname
basename(fls)
fls \%>\%
    parse_path() \%>\%
    back()

dirname(fls)
fls \%>\%
    parse_path() \%>\%
    front()
}
\seealso{
\code{\link[base]{basename}},
\code{\link[base]{dirname}}

Other sub.path.manipulation.functions: \code{\link{after}},
  \code{\link{before}}; \code{\link{index}};
  \code{\link{swap}}, \code{\link{swap_index}},
  \code{\link{swap_regex}}
}

