% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.distal.pts.R
\name{get.distal.pts}
\alias{get.distal.pts}
\title{Sample the point cloud}
\usage{
get.distal.pts(pts, min.z, plot)
}
\arguments{
\item{pts}{and object created using the function read.cloud().}

\item{min.z}{value, minimal height of points sampled.}

\item{plot}{logical, if TRUE the sampled points are plotted in respect to horizontal and vertical planes.}
}
\value{
a data.frame with sampled points coordinates.
}
\description{
Get the starting positions of paths. The output set of points includes: all points forming the crown projection convex chull (CPCH), a tree's highest point, and a set of randomly selected points. There are 100 points sampled per each meter of tree vertical extent. The probability of point selection is proportional to its distance from tree base.
}
\examples{
# read point cloud
pts <- read.cloud(file = "C:/.../point_cloud_file_name.txt", sep = " ", plot = TRUE) 
# get distal points
distal.pts <- get.distal.pts(pts, min.z = 1.3, plot = TRUE)

}
