% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent_path.R
\name{indent_path}
\alias{indent_path}
\title{Separate and Indent Path Elements}
\usage{
indent_path(path = NULL, copy2clip = is.global(2), symbol = "->")
}
\arguments{
\item{path}{A path to the file/directory.  If \code{NULL} \code{indent_path}
attempts to use \code{\link[clipr]{read_clip}} to read and fix Windows paths.}

\item{copy2clip}{logical.  If \code{TRUE} attempts to copy to the clipboard
using \code{\link[clipr]{write_clip}}.  Default copies to the clipboard if
used interactively.}

\item{symbol}{The prefix symbol to use before the path element.}
}
\description{
Separate a path into hierarchical drectories and files for indented reading.
}
\examples{
indent_path("C:/Users/Tyler/Desktop/guy")
indent_path("C:/Users/Tyler/Desktop/guy", symbol="--")
indent_path("C:/Users/Tyler/Desktop/guy", symbol=".")
indent_path("C:/Users/Tyler/Desktop/guy", symbol="")
}

