% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Extract an Index of Elements of a Path}
\usage{
index(x, inds, collapse = TRUE, ...)
}
\arguments{
\item{x}{A \code{parse_path} object.}

\item{inds}{A vector of indices to select from.}

\item{collapse}{logical.  If \code{TRUE} the path elements will be collapsed
with the \code{"/"} separator.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector (either \code{atomic} or \code{\link[base]{list}} of
\code{character} path elements) of sub-paths.
}
\description{
Extract the indices of a path.  Stands in contrast to \code{front}/\code{back}
in that with the later two the user wants the ends of the path elements (front
or bacK) whereas with \code{index} the user wants the middle elements of a path.
}
\examples{
index(list(1:8), 3)
index(list(1:8), 4:6)
index(list(1:8), c(1, 3))

library(magrittr)
files \%>\%
    parse_path() \%>\%
    index(c(2, 4:5))

files \%>\%
    parse_path() \%>\%
    index(3)
}
\seealso{
Other sub-path manipulation functions: \code{\link{before}},
  \code{\link{front}}, \code{\link{swap}}
}
