% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_url}
\alias{is_url}
\title{Is object a URL?}
\usage{
is_url(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
Logical.
\itemize{
\item TRUE: all elements in the object are URLs.
\item FALSE: all elements in the object are not URLs
}
}
\description{
Is object a URL?
}
\examples{
# The following returns TRUE
url <- "https://github.com/KoderKow/pathr"

is_url(url)

# The following returns FALSE
url <- "ww.exampl.co/this/will/fail"

try(assert_url(url))
}
