% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Ensure Number and Direction of Slash Separators Are Consistent}
\usage{
normalize(x)
}
\arguments{
\item{x}{A \code{parse_path} object.}
}
\value{
Returns an atomic vector of normalized paths.
}
\description{
Replace path separators with a single foward slash.
}
\examples{
library(magrittr)
c("C:\\\\Users\\\\Tyler\\\\AppData\\\\Local\\\\Temp\\\\Rtmp2Ll9d9", "C:/R/R-3.2.2") \%>\%
    parse_path() \%>\%
    normalize()
}

