% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_path.R
\name{parse_path}
\alias{parse_path}
\title{Parse Path(s)}
\usage{
parse_path(x)
}
\arguments{
\item{x}{A vector of paths.}
}
\value{
Returns a list of vectors of parsed paths.
}
\description{
Parses the individual directories, sub-directories, and files in a path into
individual elements.
}
\examples{
parse_path(files)
}
