% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathr.R
\name{path_basename}
\alias{path_basename}
\title{Base name of pathname path}
\usage{
path_basename(path)
}
\arguments{
\item{path}{Character vector of paths.}
}
\value{
Character vector of base names.
}
\description{
The result is the same as for \code{\link[base]{basename}},
and this function is only included here for completeness.
}
\details{
Note that this is different from the Python \code{os.path.basename}
function, Python always returns a file name or an empty string,
i.e. \code{os.path.basename('/foo/bar/')} is \code{''}. In R
\code{path_basename("/foo/bar/")} is \code{"bar"}.
}
\examples{
path_basename(c("/foo/bar", "/foo/bar/", ".", "foo/bar.c"))
}
\seealso{
\code{\link{path_dirname}}
}

