% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{path_dirname}
\alias{path_dirname}
\title{Directory name of a path}
\usage{
path_dirname(path)
}
\arguments{
\item{path}{Character vector of paths.}
}
\value{
Character vector of directory names.
}
\description{
Return the directory name of pathname path.
The result is the same as for \code{\link[base]{dirname}},
and this function is only included here for completeness.
}
\examples{
path_dirname(c("/foo/bar", "/foo/bar/", ".", "foo/bar.c", "~"))
}
\seealso{
\code{\link{path_basename}}
}

