% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{path_exists}
\alias{path_exists}
\alias{path_lexists}
\title{Test if path exists}
\usage{
path_exists(paths)

path_lexists(paths)
}
\arguments{
\item{paths}{Vector of paths to test}
}
\value{
Logical vector.
}
\description{
Returns \code{FALSE} for broken symbolic links. On some platforms, this
function may return \code{FALSE} if permission is not granted to execute
the \code{fstat} system call on the requested file, even if the
path physically exists.
}
\examples{
path_exists(c('/tmp/foobar', tempdir(), NA))
}

