% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_abs.R
\name{path_is_abs}
\alias{path_is_abs}
\title{Test for absolute paths}
\usage{
path_is_abs(path)
}
\arguments{
\item{path}{Character vector of absolute paths.}
}
\value{
Logical vector.
}
\description{
On Unix, that means it begins with a slash, on Windows that it
begins with a (back)slash after chopping off a potential drive
letter.  On Windows, UNC paths are considered to be absolute paths.
}
\examples{
path_is_abs(c("/foo/bar/", "./relative", "../me/too", "and/me"))
}

