% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_directory.R
\name{path_is_directory}
\alias{path_is_directory}
\title{Test if a path is a directory}
\usage{
path_is_directory(path, missing_is_not_directory = FALSE)
}
\arguments{
\item{path}{A vector of paths}

\item{missing_is_not_directory}{Flag to indicate a return value of
\code{FALSE} for missing files (not for missing values in
\code{path}).  By default missing files have an value of
\code{NA} which matches R's NA semantics, but not Python's
\code{os.path.isdir} function.  Pass \code{TRUE} in here to set
missing files to \code{FALSE} (i.e., a missing path is
\emph{not} a directory, rather than having an unknown state of
directory-ness).}
}
\description{
Test if a path is a directory.  This follows symbolic links, so
both \code{\link{path_is_link}} and \code{path_is_directory} can
be \code{TRUE} for the same oath on systems that support symlinks
}
\examples{
path_is_directory(".") # TRUE
path_is_directory("missing file") # NA
}

