% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_link.R
\name{path_is_link}
\alias{path_is_link}
\title{Test for symlink}
\usage{
path_is_link(path, absent_is_false = TRUE)
}
\arguments{
\item{path}{Paths to test}
}
\value{
Logical vector.
}
\description{
Always false if symbolic links are not supported by the platform.
Returns \code{FALSE} for non-existent paths.
}
\examples{
path_is_link(c(tempdir(), "/tmp/foo", NA_character_))
}

