% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{path_norm}
\alias{path_norm}
\title{Normalize a pathname}
\usage{
path_norm(path)
}
\arguments{
\item{path}{Character vector of paths.}
}
\value{
Character vector of normalized paths.
}
\description{
Normalize a pathname by collapsing redundant separators and
up-level references so that \code{A//B}, \code{A/B/}, \code{A/./B} and
\code{A/foo/../B} all become \code{A/B}. This string manipulation may
change the meaning of a path that contains symbolic links.
}
\details{
On Windows, it converts forward slashes to backward slashes.
}
\examples{
path_norm(c("A//B", "A/B/", "A/./B", "A/foo/../B"))
}

