% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rel.R
\name{path_rel}
\alias{path_rel}
\title{Relative filepath from a path to another}
\usage{
path_rel(path, start = ".")
}
\arguments{
\item{path}{Character vector of paths.}

\item{start}{Start directory, character scalar. Defaults to
the current working directory.}
}
\description{
Return a relative filepath to another path, either from the current
directory or from an optional start directory. This is a path
computation: the filesystem is not accessed to confirm the
existence or nature of path or start.
}
\examples{
path_rel("./bar/foo/bar", "/")
path_rel("./bar/foo/bar", "bar/baz/ize")
path_rel("./bar/foo/bar", "../../")
}

