% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_drive.R
\name{path_split_drive}
\alias{path_split_drive}
\title{Split a pathname into drive/UNC sharepoint and relative path specifiers}
\usage{
path_split_drive(path)
}
\arguments{
\item{path}{Character vector.}
}
\value{
List of character vectors of length two. The first elements
  contain the drive letters or UNC paths, the second elements the
  relative paths.
}
\description{
Returns a 2 element list with character vectors: drive/UNC and
path, either part may contain empty strings ("") but both will
have the same length as the input.
}
\details{
It always holds that \code{paste0(output[1], output[2]) == input}
(note that this means that the path part is an \emph{absolute}
path, including a leading slash, when a drive or UNC mountpoint is
present).

If the path contained a drive letter, the drive/UNC part will contain
everything up to and including the colon. E.g.
\code{path_split_drive("c:/dir")} returns
\code{list(drive = "c:", path = "/dir")}.

If the path contained a UNC path, the drive/UNC will contain the host
name and share up to but not including the fourth directory separator
character. E.g. \code{path_split_drive("//host/computer/dir")} returns
\code{list(drive = "//host/computer", path = "/dir")}.

Paths cannot contain both a drive letter and a UNC path.

Note that this differs from the Python \code{os.path} module in
that UNC parts are considered drives here.
}
\examples{
# path_split_drive(c("c:/Rtools/bin", "\\\\\\\\machine\\\\share\\\\folder"))
}

