% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathr-package.R
\docType{package}
\name{pathr-package}
\alias{pathr}
\alias{pathr-package}
\title{pathr: Work with URL and file paths}
\description{
The goal of pathr is to simplify the process of string manipulations for URLs and file paths. This uses modified functions from the {httr} to create opinionated functions that allow working with URLs and paths easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/pathr}
  \item \url{https://koderkow.github.io/pathr/}
  \item Report bugs at \url{https://github.com/KoderKow/pathr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
