% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathr.R
\name{pathr}
\alias{pathr}
\title{Path finder for mazes}
\usage{
pathr(x_max = 15, y_max = 15, col_width = 1, border_pad = 1)
}
\arguments{
\item{x_max}{Maximum canvas width of maze for \eqn{x}-axis.}

\item{y_max}{Maximum canvas width of maze for \eqn{y}-axis.}

\item{col_width}{Size of columns for canvas. Rounds to nearest whole number.}

\item{border_pad}{Uniform margin width for canvas.}
}
\description{
An example algorithm for random maze generation.
}
