% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathr_to_df.R
\name{pathr_to_df}
\alias{pathr_to_df}
\title{Turn pathr object to a data.frame}
\usage{
pathr_to_df(x)
}
\arguments{
\item{x}{A pathr object.}
}
\value{
A data.frame with as many rows as there were pathr objects.
}
\description{
Turn a single of list of pathr objects in a data.frame for analysis.
}
\examples{
urls <- c(
  "https://github.com/KoderKow/pathr1",
  "https://github.com/KoderKow/pathr2"
)

x <- parse_url(urls)

d <- pathr_to_df(x)

d
}
