% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.metrics.R
\name{paths.metrics}
\alias{paths.metrics}
\title{Tree-level metrics of branching orientation and path detection rate}
\usage{
paths.metrics(distal.pts, paths.smooth, plot = TRUE)
}
\arguments{
\item{distal.pts}{an object created using the function get.distal.pts().}

\item{paths.smooth}{an object created using the function smooth.paths().}

\item{plot}{logical, if TRUE the histogram of segment angles is plotted.}
}
\value{
metrics as a data.frame.
}
\description{
Outputs several tree-level metrics:
detect.full (full paths detection rate),
detect.all (full and partial paths detection rate),
plagio.full (percent of full plagiotropic paths),
plagio.all (percent of full and partial plagiotropic paths).
}
\examples{
# read point cloud
pts <- read.cloud(file = "C:/.../point_cloud_file_name.txt", sep = " ", plot = TRUE) 
# get distal points
distal.pts <- get.distal.pts(pts, min.z = 1.3, plot = TRUE)
# get paths
paths <- find.paths(pts, distal.pts, proximal.z = 0.0, min.range  = .02, max.range  = .08)
# normalize segment length
paths.smooth <- smooth.paths(paths, length.f = .10)
# get tree-level metrics
paths.metrics(distal.pts, paths.smooth, plot = TRUE)

}
