% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indent_path.R
\name{print.indent_path}
\alias{print.indent_path}
\title{Prints an indent_path object}
\usage{
\method{print}{indent_path}(x, ...)
}
\arguments{
\item{x}{An \code{indent_path} object.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}.}
}
\description{
Prints an indent_path object.
}

