% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_class.R
\name{print.manipulate}
\alias{print.manipulate}
\title{Prints a manipulate Object}
\usage{
\method{print}{manipulate}(x, ...)
}
\arguments{
\item{x}{A manipulate object}

\item{\ldots}{ignored.}
}
\description{
Prints a manipulate object
}

