% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_path.R
\name{print.parse_path}
\alias{print.parse_path}
\title{Prints a parse_path Object}
\usage{
\method{print}{parse_path}(x, ...)
}
\arguments{
\item{x}{A parse_path object}

\item{\ldots}{ignored.}
}
\description{
Prints a parse_path object
}

