% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print.tree}
\alias{print.tree}
\title{Prints a tree object}
\usage{
\method{print}{tree}(x, ...)
}
\arguments{
\item{x}{A \code{tree} object.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}.}
}
\description{
Prints a tree object.
}

