% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.paths.R
\name{project.paths}
\alias{project.paths}
\title{Plot classified paths}
\usage{
project.paths(pts, distal.pts, paths.smooth)
}
\arguments{
\item{pts}{and object created using the function read.cloud().}

\item{distal.pts}{an object created using the function get.distal.pts().}

\item{paths.smooth}{and object created using the function smooth.paths().}
}
\value{
no data are returned.
}
\description{
The paths (classified as plagiotropic or orthotropic) are projected onto vertical and horizontal planes.
}
\examples{
# read point cloud
pts <- read.cloud(file = "C:/.../point_cloud_file_name.txt", sep = " ", plot = TRUE) 
# get distal points
distal.pts <- get.distal.pts(pts, min.z = 1.3, plot = TRUE)
# get paths
paths <- find.paths(pts, distal.pts, proximal.z = 0.0, min.range  = .02, max.range  = .08)
# normalize segment length
paths.smooth <- smooth.paths(paths, length.f = .10)
# plot
project.paths(pts, distal.pts, paths.smooth)

}
