% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.cloud.R
\name{read.cloud}
\alias{read.cloud}
\title{Read TLS point cloud}
\usage{
read.cloud(file, sep, plot)
}
\arguments{
\item{file}{point cloud file directory and name.}

\item{sep}{character, separator.}

\item{plot}{logical, if TRUE the point cloud is plotted.}
}
\value{
Point cloud as a data.frame.
}
\description{
This function reads TLS point cloud into R,
the allowed formats include .txt, .csv, .pts (containing three columns, with the x, y and z point coordinates),
row names and column names are not allowed.
}
