% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pathr.R
\name{remove_pathr}
\alias{remove_pathr}
\title{Remove pathr class}
\usage{
remove_pathr(x)
}
\arguments{
\item{x}{Object to remove pathr class from.}
}
\value{
This function handles two things
\itemize{
\item Single object: The object, x, returned with \code{\link[=unclass]{unclass()}} called if the class pathr existed
\item List of objects: The list of objects, x, returned with \code{\link[=unclass]{unclass()}} called on each object if the class of all items in the list of objects are pathr
}
}
\description{
Simplifies removing the pathr class for situations where the custom class, pathr, is causing issues with workflows.
}
\examples{
url <- "https://github.com/KoderKow/pathr"

x <- parse_url(url)

class(x)

y <- remove_pathr(x)

class(y)
}
