% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_ui.R
\name{selection_ui}
\alias{selection_ui}
\title{Select from a vector}
\usage{
selection_ui(id, selection_label = "Document:")
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{selection_label}{Character. Label to indicate what is being selected.}
}
\value{
The selected element of the supplied vector.
}
\description{
Module facilitating the selection of a specific value in a vector by cycling through elements of a vector.
}
\author{
Nicolas Mangin
}
