% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.paths.R
\name{smooth.paths}
\alias{smooth.paths}
\title{Smooth paths}
\usage{
smooth.paths(paths, length.f)
}
\arguments{
\item{paths}{an object created using the function find.paths().}

\item{length.f}{value, new segment length (should be larger than max.range in find.paths()).}
}
\value{
a list including smoothed paths.
}
\description{
Normalize segment length within the paths.
}
\examples{
# read point cloud
pts <- read.cloud(file = "C:/.../point_cloud_file_name.txt", sep = " ", plot = TRUE) 
# get distal points
distal.pts <- get.distal.pts(pts, min.z = 1.3, plot = TRUE)
# get paths
paths <- find.paths(pts, distal.pts, proximal.z = 0.0, min.range  = .02, max.range  = .08)
# normalize segment length
paths.smooth <- smooth.paths(paths, length.f = .10)

}
