% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\alias{swap_index}
\alias{swap_regex}
\title{Replace Elements of a Path}
\usage{
swap(x, inds, replacement, pattern = NULL, collapse = TRUE, ...)

swap_index(x, inds, replacement, collapse = TRUE, ...)

swap_regex(x, pattern, replacement, collapse = TRUE, ...)
}
\arguments{
\item{x}{A \code{parse_path} object.}

\item{inds}{Indices to replace.}

\item{replacement}{A value to replace the indentified indices with.}

\item{pattern}{A character string containing a regular expression (or
character string for \code{fixed = TRUE}) to be matched in the given character
vector. Note that under the hood \code{\link[base]{grep}} is used to locate
the indices of path elements that match the regex condition.  Only the first
match will be used.  The argument \code{perl = TRUE} is hard coded.}

\item{collapse}{logical.  If \code{TRUE} the path elements will be collapsed
with the \code{"/"} separator.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector (either \code{atomic} or \code{\link[base]{list}} of
\code{character} path elements) of swapped paths.
}
\description{
Replace the elements of a path. Note either a index (use \code{inds}) or regex
(use \code{pattern}) based approach is used to identify the indices to replace.
The more specific \code{swap_index} or \code{swap_regex} can be used to be
more targetted, transparent, pipeable, and efficient.
}
\examples{
swap(list(1:8), inds = c(2, 4), "REPLACED VALUE")
swap(list(1:8), pattern = "[238]", replacement = "REPLACED VALUE")

library(magrittr)
files \%>\%
    parse_path() \%>\%
    swap(1, "MyFile")

files \%>\%
    parse_path() \%>\%
    swap_index(3, "textMining")

files \%>\%
    parse_path() \%>\%
    swap_regex("^qdap$", "textMining")

files \%>\%
    parse_path() \%>\%
    swap_regex("\\\\.R$", "function.R")

files \%>\%
    parse_path() \%>\%
    swap_regex("^(P|R)", "oops")
}
\seealso{
Other sub-path manipulation functions: \code{\link{before}},
  \code{\link{front}}, \code{\link{index}}
}
