% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Generate a Directory Tree}
\usage{
tree(path = ".", include.files = TRUE, all.files = TRUE,
  use.data.tree = FALSE, out = NULL, additional = NULL,
  copy2clip = FALSE)
}
\arguments{
\item{path}{A path to the directory.}

\item{include.files}{logical.  If \code{TRUE} files and drectories will be
included.}

\item{all.files}{logical.   If \code{FALSE}, only the names of visible files
are returned. If \code{TRUE}, all file names will be returned.}

\item{out}{An optional outfile (typically .txt) to print to.}

\item{additional}{Additonal arguments passed to
href{https://www.microsoft.com/resources/documentation/windows/xp/all/proddocs/en-us/tree.mspx?mfr=true}{\code{tree}}
(Windows) or \href{http://www.computerhope.com/unix/tree.htm}{tree} (Unix).}

\item{copy2clip}{logical.  If \code{TRUE} attempts to copy to the clipboard
via \code{\link[clipr]{write_clip}}.}
}
\description{
Given a directory, generate a directory tree diagram.
}
\note{
The \pkg{tree} program must be installed to use \code{use.data.tree = FALSE}.
}
\examples{
tree(system.file("", package = "base"), use.data.tree = TRUE)
\dontrun{
tree()
tree(.libPaths())
}
}
