% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{url_regex}
\alias{url_regex}
\title{Regex to validate URLs}
\format{
Character. Length 1.
}
\source{
\url{https://stackoverflow.com/a/17714711}
}
\usage{
url_regex
}
\description{
Regex to validate URLs
}
\concept{Data}
\keyword{datasets}
