% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_fix.R
\name{win_fix}
\alias{win_fix}
\title{Fix Single Windows Backslashes}
\usage{
win_fix(clip = TRUE, quote = TRUE, copy2clip = interactive())
}
\arguments{
\item{clip}{logical.  If \code{TRUE} the path is copied from the clipboard
via the \pkg{clipr} package.}

\item{quote}{logical.  If \code{TRUE} the returned path is wrapped with quotes.}

\item{copy2clip}{logical.  If \code{TRUE} the returned path is copied to the
clipboard.}
}
\value{
Returns an R friendly path.
}
\description{
Fix single Windows based backslashes in paths with a single foward slash.  By
defaukt copying a path from a Windows machine has R unfriendly single
backslashes.  This function copies a single Windows path from the clipboard
maes the necessary single foward slash correction.
}
\examples{
## copy this line to clipboard...
## ~\\Packages\\qdap\\R\\cm_code.overlap.R
## Now run...
## win_fix()
}

