% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abspath.R
\name{path_abs}
\alias{path_abs}
\title{Normalized, absolutized paths}
\usage{
path_abs(path)
}
\arguments{
\item{path}{Character vector.}
}
\value{
Character vector.
}
\description{
Return a normalized absolutized version of the pathname path. On
most platforms, this is equivalent to calling the function
\code{\link{path_norm}} as follows:
\code{path_norm(file.path(getwd(), path))}.
}
\examples{
path_abs(c("foo/bar", "foo/..", "/root/dir"))
}

