% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdir.R
\name{path_is_directory}
\alias{path_is_directory}
\title{Test if a path is a directory}
\usage{
path_is_directory(path)
}
\arguments{
\item{path}{A vector of paths}
}
\description{
Test if a path is a directory.  This follows symbolic links, so
both \code{\path{path_is_link}} and \code{path_is_directory} can
be \code{TRUE} for the same oath on systems that support symlinks
}
\examples{
path_is_directory(".") # TRUE
path_is_directory("missing file") # NA
}

