% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathr-package.R
\docType{package}
\name{pathr-package}
\alias{pathr}
\alias{pathr-package}
\title{pathr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/pathr}
  \item \url{https://koderkow.github.io/pathr/}
  \item Report bugs at \url{https://github.com/KoderKow/pathr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
